COMP: FIx warning. Committed on the Free edition of March Hare Software CVSNT Server. Upgrade to CVS Suite for more features and support: http://march-hare.com/cvsnt/

+++ VTK/Rendering/vtkLabelHierarchy.cxx
@@ -1565,7 +1565,7 @@ void vtkLabelHierarchy::implementation::DropAnchor( vtkIdType anchor )
   if(mapIter == this->coordMap.end())
     {
     vtkstd::pair<int, vtkstd::set<vtkIdType> > Pair;
-    Pair.first = curs.level();
+    Pair.first = static_cast<int>(curs.level());
     Pair.second.insert(anchor);
     this->coordMap[coord] = Pair;
     }
